migrate to new schema architecture
